+++ src/Commands/Enums/MultiGeoExperienceMode.cs
@@ -1,0 +1,18 @@
+namespace PnP.PowerShell.Commands.Enums
+{
+	/// <summary>
+	/// Specifies the SharePoint Online multi-geo experience mode.
+	/// </summary>
+	public enum MultiGeoExperienceMode
+	{
+		/// <summary>
+		/// SharePoint Online Multi-Geo experience mode.
+		/// </summary>
+		SPO = 0,
+
+		/// <summary>
+		/// OneDrive Multi-Geo experience mode.
+		/// </summary>
+		ODB = 1
+	}
+}